Move layout to components
Current approach breaks ECS principles because a core component es dependent on an engine package
