+++ internal/core/components/maze.go
@@ -1,11 +1,7 @@
 package components
 
-import (
-	"github.com/juanancid/maze-adventure/internal/engine/mazebuilder"
-)
-
 type Maze struct {
-	Layout     mazebuilder.Layout
+	Layout     Layout
 	CellWidth  int
 	CellHeight int
 }
